fix(gui): create some of widgets by function (not struct) to fix bug of "initial ui is not freshed"
In preview code, some checkboxes and enties are created by struct, which can cause the initial ui
not updated. e.g. If the initial state of a checkbox is checked and create checkbox by struct, then
the ui shown on app is unchecked after app launch. We must create checkbox by function
'widget.NewCheck' and set is checked.