fix(client-ui): add customized Info.plist for mac client-ui app, which make app termination gracefully
When the client connection with wssocks server is lost (e.g. wifi off, or no activity in a
long time), the ListenIncomeMsg function or HeartBeat does not return with error (In fact,
ListenIncomeMsg does not return).
In practice, we should terminate this app and restart it. But when clicking the 'stop' button,
it can make ListenIncomeMsg return. It just show 'stopping' text on the button endlessly.
When we try to close the app by quit menu or ⌘+ Q, it still does not have any response.
We must force kill it in Activity Monitor.

In this commit, by adding a customized Info.plist for mac client-ui app, which include
`NSSupportsSuddenTermination` field, then we can terminate app gracefully.

related issue: ref #4