feat(gui): wait and show fyne-based client error message after the client is started
after the fyne-based client is started, we call api handles.Wait() to wait error and show error
message dialog if there is some error while running (e.g. connection lost or listern port already in
used).
This commit is similar as PR #12. This is for the fyne-based client, while #12 is for swiftui client.